Importantly, recognised that the calling of the player deaths update action should be a callback from the update of the collision matrix update, conditional on whether anything was updated (just like the dispatch, but seperate from the dispatch)
I feel like some poor decisions are being made elsewhere, like the
inconsistency in passing values back on event emissions and the fuzzy
layers which are appearing. However, once the control panel and
scoreboard are connected up to this and the working demo can be replaced
in the source folder I can review all the design and the tests.